requirements.txt
claude-agent-sdk
litellm[proxy]
python-dotenv
